move mainutils to main.go
